requirements.txt
absl-py==0.11.0
actionlib==1.12.1
angles==1.9.12
astor==0.8.1
astunparse==1.6.3
attrs==20.3.0
bondpy==1.8.5
braceexpand==0.1.6
cached-property==1.5.2
cachetools==4.2.1
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.29
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cmake==3.18.4.post1
Connectivity==0.0.0
controller-manager==0.18.3
controller-manager-msgs==0.18.3
cv-bridge==1.13.0
cycler==0.10.0
Cython==0.29.24
decorator==4.4.2
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
dynamic-reconfigure==1.6.3
dynamixel-sdk==3.7.51
filelock==3.3.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
gitdb==4.0.5
GitPython==3.1.12
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gym==0.10.9
h5py==2.10.0
huggingface-hub==0.0.19
idna==2.10
ifcfg==0.21
image-geometry==1.13.0
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib-metadata==3.4.0
interactive-markers==1.11.5
joblib==1.0.0
joint-state-publisher==1.12.15
joint-state-publisher-gui==1.12.15
kdl-parser-py==1.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
laser-geometry==1.6.7
llvmlite==0.35.0
lmdb==1.0.0
Markdown==3.3.3
matplotlib==3.3.4
message-filters==1.14.11
mock==4.0.3
moveit-commander==1.0.7
moveit-ros-planning-interface==1.0.7
moveit-ros-visualization==1.0.7
moviepy==1.0.3
msgpack==1.0.2
networkx==2.5
numba==0.52.0
numpy==1.19.5
numpy-quaternion==2020.11.2.17.0.49
oauthlib==3.1.0
objectio==0.2.29
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==21.0
pickle5==0.0.11
Pillow==7.2.0
plyfile==0.7.2
proglog==0.1.9
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-qt-binding==0.4.3
PyWavelets==1.1.1
PyYAML==5.4.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
regex==2021.10.8
requests==2.25.1
requests-oauthlib==1.3.0
resource-retriever==1.12.6
rosbag==1.14.11
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.11
roslaunch==1.14.11
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.11
rosmake==1.14.9
rosmaster==1.14.11
rosmsg==1.14.11
rosnode==1.14.11
rosparam==1.14.11
rospy==1.14.11
rosservice==1.14.11
rostest==1.14.11
rostopic==1.14.11
rosunit==1.14.9
roswtf==1.14.11
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.10
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rsa==4.7
rviz==1.13.17
sacremoses==0.0.46
scikit-fmm==0.0.7
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.6.0
sensor-msgs==1.12.8
simplejson==3.17.2
six==1.15.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
smmap==3.0.5
srdfdom==0.5.2
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
threadpoolctl==2.1.0
tifffile==2021.1.14
tokenizers==0.10.3
topic-tools==1.14.11
torch==1.7.1
torchgeometry==0.1.2
torchvision==0.8.2
tqdm==4.56.0
transformers==4.11.3
typer==0.3.2
typing-extensions==3.7.4.3
urdfdom-py==0.4.5
urllib3==1.26.3
webdataset==0.1.40
Werkzeug==1.0.1
wrapt==1.12.1
xacro==1.13.10
yacs==0.1.8
zipp==3.4.0